A shit ton of work on config and testing Added getConfig(value, expectedType) that automatically converts values Removed Objenesis, no longer needed Added simple autocomplete to "value" field in /setproperty Changed ConfigProperty to now store KClass instead of class names for much easier converting Added GameTest's for config Updated Location to store Int's again Added Utils.Colors to store useful color values in an enum instead of using class members, store colors as strings instead of Int's now for easier configuration (might change in the future)
Took 3 hours 25 minutes

+++ src/main/kotlin/net/vannername/qol/utils/Location.kt
@@ -9,15 +9,16 @@ import kotlin.math.pow
 import kotlin.math.sqrt
 
 class Location(
-    val x: Double, val y: Double, val z: Double, val worldName: String
+    val x: Int, val y: Int, val z: Int, val worldName: String
 )
 {
-    constructor(x: Int, y: Int, z: Int, worldName: String) : this(x.toDouble(), y.toDouble(), z.toDouble(), worldName)
+    constructor(x: Double, y: Double, z: Double, worldName: String) : this(x.toInt(), y.toInt(), z.toInt(), worldName)
+
     fun getWorld(server: MinecraftServer): ServerWorld {
         return server.getWorld(RegistryKey.of(RegistryKeys.WORLD, Identifier(worldName)))!!
     }
 
     fun getDistance(l: Location): Double {
-        return sqrt((x - l.x).pow(2) + (y - l.y).pow(2) + (z - l.z).pow(2))
+        return sqrt((x - l.x).toDouble().pow(2) + (y - l.y).toDouble().pow(2) + (z - l.z).toDouble().pow(2))
     }
 }
